backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# Google APIs
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# LLM Providers
anthropic==0.25.1
openai==1.58.1
google-generativeai==0.8.3

# Background Jobs
celery==5.3.6
redis==5.0.1

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
email-validator==2.1.0
python-dateutil==2.8.2

# CORS
fastapi-cors==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
